Move recordings management to own process and ensure db multiprocess access
