+++ frigate/record/util.py
@@ -1,0 +1,21 @@
+"""Recordings Utilities."""
+
+import os
+
+from frigate.const import RECORD_DIR
+
+
+def remove_empty_directories(directory):
+    # list all directories recursively and sort them by path,
+    # longest first
+    paths = sorted(
+        [x[0] for x in os.walk(RECORD_DIR)],
+        key=lambda p: len(str(p)),
+        reverse=True,
+    )
+    for path in paths:
+        # don't delete the parent
+        if path == RECORD_DIR:
+            continue
+        if len(os.listdir(path)) == 0:
+            os.rmdir(path)
